Actalizando capa de negocio

+++ PetCenter.BusinessEntities/MonitoreoBE.cs
@@ -9,7 +9,7 @@ namespace PetCenter.BusinessEntities
     public class MonitoreoBE
     {
         public int codigo { get; set; }
-        public int lugarHospedaje { get; set; }
+        public long lugarHospedaje { get; set; }
         public int mascota { get; set; }
         public string observaciones { get; set; }
         public string fechaRegistro { get; set; }
